refactor: extract admin import and sync endpoints into dedicated router
